Autenticação no fron-end

+++ server/src/routes/memories.ts
@@ -1,12 +1,21 @@
 import { FastifyInstance } from "fastify";
 import { prisma } from "../lib/prisma";
 import { z } from 'zod';
+import { request } from "http";
 
 
 export async function memoriesRoutes(app: FastifyInstance) {
 
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async (req) => {
+        await req.jwtVerify()
+    })
+
+    app.get('/memories', async (req) => {
+
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: req.user.sub
+            },
             orderBy: {
                 createdAt: 'asc'
             }
@@ -21,13 +30,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
 
 
-    app.get('/memories/:id', async (req) => {
+    app.get('/memories/:id', async (req, res) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
 
         const { id } = paramsSchema.parse(req.params);
         const memory = await prisma.memory.findUniqueOrThrow({ where: { id } });
+
+        if(!memory.isPublic && memory.userId != req.user.sub) {
+            return res.status(401);
+        }
 
         return memory
     });
@@ -46,14 +59,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content, 
                 coverUrl,
                 isPublic,
-                userId: 'd0f9ba3e-6e2b-4f2d-b472-04b804abff50'
+                userId: req.user.sub
             }
         });
         return memory;
     });
 
 
-    app.put('/memories/:id', async (req) => {
+    app.put('/memories/:id', async (req, res) => {
+
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
@@ -66,7 +80,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
         const { id } = paramsSchema.parse(req.params);
         const { content, coverUrl, isPublic } = bodySchema.parse(req.body);
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({ where: { id } });
+
+        if(!memory.isPublic && memory.userId != req.user.sub) {
+            return res.status(401);
+        }
+
+        memory = await prisma.memory.update({
             where: {
                 id
             },
@@ -81,12 +101,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
 
 
-    app.delete('/memories/:id', async (req) => {
+    app.delete('/memories/:id', async (req, res) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
 
         const { id } = paramsSchema.parse(req.params);
-        const memory = await prisma.memory.delete({ where: { id } });
+
+        let memory = await prisma.memory.findUniqueOrThrow({ where: { id } });
+        
+        if(!memory.isPublic && memory.userId != req.user.sub) {
+            return res.status(401);
+        }
+
+        memory = await prisma.memory.delete({ where: { id } });
     });
 }
